backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
requests
numpy
opencv-python
pydicom
torch
transformers
segment-anything
python-dotenv